% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{total}
\alias{total}
\title{Adding Numbers}
\usage{
total(x)
}
\arguments{
\item{x}{Input vector of numbers}
}
\value{
Value of sum
}
\description{
The purpose of this function is to find the sum of a vector of numbers.
}
\examples{
mytotal <- total(c(1,2,3,4,5))
mytotal <- total(x)
}

